+++ tests/Client/Exception/HandlerNotFoundTest.php
@@ -5,15 +5,14 @@ declare(strict_types=1);
 namespace DigipolisGent\Tests\API\Client\Exception;
 
 use DigipolisGent\API\Client\Exception\HandlerNotFound;
+use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 use Prophecy\PhpUnit\ProphecyTrait;
 use Psr\Http\Message\RequestInterface;
 
-/**
- * @covers \DigipolisGent\API\Client\Exception\HandlerNotFound
- */
-class HandlerNotFoundTest extends TestCase
+#[CoversClass(HandlerNotFound::class)]
+final class HandlerNotFoundTest extends TestCase
 {
     use ProphecyTrait;