+++ tests/Client/Exception/InvalidResponseTest.php
@@ -10,6 +10,7 @@ use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 use Prophecy\PhpUnit\ProphecyTrait;
 use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\StreamInterface;
 
 #[CoversClass(InvalidResponse::class)]
 final class InvalidResponseTest extends TestCase
@@ -25,8 +26,11 @@ final class InvalidResponseTest extends TestCase
         $data = json_encode(['value' => uniqid('', true)], JSON_THROW_ON_ERROR);
         $statusCode = random_int(200, 500);
 
+        $stream = $this->prophesize(StreamInterface::class);
+        $stream->__toString()->willReturn($data);
+
         $response = $this->prophesize(ResponseInterface::class);
-        $response->getBody()->willReturn($data);
+        $response->getBody()->willReturn($stream->reveal());
         $response->getStatusCode()->willReturn($statusCode);
 
         $exception = InvalidResponse::fromResponse($response->reveal());