+++ tests/Logger/ResponseLogTest.php
@@ -10,6 +10,7 @@ use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 use Prophecy\PhpUnit\ProphecyTrait;
 use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\StreamInterface;
 
 #[CoversClass(ResponseLog::class)]
 final class ResponseLogTest extends TestCase
@@ -22,10 +23,13 @@ final class ResponseLogTest extends TestCase
     #[Test]
     public function castToStringHasAllDetails(): void
     {
+        $stream = $this->prophesize(StreamInterface::class);
+        $stream->__toString()->willReturn('bodyTest');
+
         $response = $this->prophesize(ResponseInterface::class);
         $response->getStatusCode()->willReturn(400);
         $response->getHeaders()->willReturn(['test' => 'foo']);
-        $response->getBody()->willReturn('bodyTest');
+        $response->getBody()->willReturn($stream->reveal());
 
         $logItem = new ResponseLog($response->reveal());